Add cast profiles

+++ fetch/fetch.rb
@@ -70,9 +70,57 @@ list.each do |film|
         url = "https://image.tmdb.org/t/p/w500#{image_slug}"
         response = Net::HTTP.get(URI(url))
         puts "Writing #{response.length} bytes to #{path}"
-        File.write(full_path, response)    
+        File.write(full_path, response)
       end
       film[type] = path
+    end
+  end
+end
+
+# 5. Collect all unique people from film credits and fetch their details
+people_data_dir = "people"
+people_images_dir = "../content/bill/films/profiles"
+Dir.mkdir(people_data_dir) unless Dir.exist?(people_data_dir)
+Dir.mkdir(people_images_dir) unless Dir.exist?(people_images_dir)
+
+people = {}
+list.each do |film|
+  filename = "../_data/films/#{film["slug"]}.json"
+  next unless File.exist?(filename)
+  data = JSON.parse(IO.read(filename))
+  credits = data["credits"] || {}
+
+  (credits["cast"] || []).sort_by { |p| p["order"] || 999 }.first(12).each do |person|
+    people[person["id"]] ||= { "name" => person["name"], "profile_path" => person["profile_path"] }
+  end
+
+  (credits["crew"] || []).select { |p| p["job"] == "Director" }.each do |person|
+    people[person["id"]] ||= { "name" => person["name"], "profile_path" => person["profile_path"] }
+  end
+end
+
+puts "Found #{people.size} unique people across all films"
+
+# Fetch person details from TMDB
+people.each do |person_id, person|
+  filename = "#{people_data_dir}/#{person_id}.json"
+  if !File.exist?(filename) || ARGV[0] == "--force"
+    puts "Fetching details for #{person["name"]} (#{person_id})"
+    response = get_api_response(endpoint: "person/#{person_id}")
+    data = JSON.parse(response.read_body)
+    IO.write(filename, JSON.pretty_generate(data))
+  end
+end
+
+# Download profile images
+people.each do |person_id, person|
+  if profile_path = person["profile_path"]
+    full_path = "#{people_images_dir}/#{person_id}.jpg"
+    if !File.exist?(full_path)
+      url = "https://image.tmdb.org/t/p/w500#{profile_path}"
+      response = Net::HTTP.get(URI(url))
+      puts "Writing #{response.length} bytes to profiles/#{person_id}.jpg"
+      File.write(full_path, response)
     end
   end
 end
